requirements.txt
pydantic>=2.0.0
aiohttp>=3.8.0
jinja2>=3.1.0
pyyaml>=6.0
click>=8.0.0
openai>=1.0.0
anthropic>=0.8.0
python-dotenv>=1.0.0
tqdm>=4.64.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1